Adds ESlint and Stylelint
