+++ .stylelintrc.json
@@ -1,0 +1,6 @@
+{
+    "extends": "stylelint-config-sass-guidelines",
+    "rules": {
+      "indentation": null
+    }
+}
