+++ package.json
@@ -6,7 +6,8 @@
     "clean": "rimraf ./dist/ && rimraf ./.parcel-cache/",
     "dev": "npm run clean && parcel src/index.html",
     "build": "npm run clean && parcel build src/index.html",
-    "start": "npm run clean && parcel build src/index.html && node server.js"
+    "start": "npm run clean && parcel build src/index.html && node server.js",
+    "lint": "eslint ./src && npx stylelint **/*.sass "
   },
   "keywords": [],
   "author": "",
@@ -14,6 +15,8 @@
   "devDependencies": {
     "@anikin/parcel-transformer-handlebars-precompile": "1.0.1",
     "@parcel/transformer-sass": "2.7.0",
+    "@typescript-eslint/eslint-plugin": "5.36.2",
+    "@typescript-eslint/parser": "5.36.2",
     "eslint": "8.23.0",
     "express": "4.18.1",
     "handlebars": "4.7.7",
@@ -21,6 +24,7 @@
     "parcel": "2.7.0",
     "rimraf": "3.0.2",
     "stylelint": "14.11.0",
+    "stylelint-config-sass-guidelines": "9.0.1",
     "typescript": "4.8.3"
   },
   "dependenciesComments": {
