+++ src/pages/options/options.ts
@@ -3,6 +3,8 @@ import Block from '../../utils/Block';
 import { OptionsRow, OptionsRowProps } from '../../components/options_row';
 import { Button } from '../../components/button';
 import { Arrow } from '../../components/arrow';
+import animateClick from '../../utils/animateClick';
+import { simplerouter } from '../../utils/simplerouter';
 
 export const userdata: { [index: string]: string; } = {
     email: 'pochta@yandex.ru',
@@ -34,13 +36,6 @@ interface OptionsPageProps {
     titles: Array<object>;
     titles__editinfo: Array<object>;
     titles_editpassword: Array<object>;
-    arrow: Arrow;
-    buttonsavepassword: Button;
-    buttonavatarupload: Button;
-    buttonsaveinfo: Button;
-    buttoneditinfo: Button;
-    buttonchangepassword: Button;
-    buttonbacktochats: Button;
 }
     
 export class OptionsPage extends Block<OptionsPageProps> {
@@ -50,6 +45,116 @@ export class OptionsPage extends Block<OptionsPageProps> {
         this.element!.classList.add("options")
     }
     init() {
+        this.children.arrow = new Arrow({ flip: "true" });
+
+        this.children.buttonbacktochats = new Button({
+            label: "Back to chats",
+            added_class: ["options__button", "options__button__backtochats"],
+            type: "button",
+            shape_dir: "left",
+            bgshape: true,
+            events: {
+                click: () => { 
+                    animateClick(this.children.buttonbacktochats.element);
+                    setTimeout(() =>  {
+                        simplerouter.temp()
+                    }, 400);
+                }
+            }
+        });
+        this.children.buttonchangepassword = new Button({
+            label: "Change password",
+            added_class: ["options__button", "options__button__changepassword"],
+            type: "button",
+            shape_dir: "right",
+            bgshape: true,
+            events: {
+              click: () => { 
+                animateClick(this.children.buttonchangepassword.element);
+                setTimeout(() =>  {
+                    for(const profile of document.querySelectorAll<HTMLElement>('.options__profile'))
+                        profile.style.display = 'none';
+                    for(const editpassword of document.querySelectorAll<HTMLElement>('.options__editpassword'))
+                        editpassword.style.display = 'block';
+                    }, 400);
+                }
+            }
+        });
+        this.children.buttoneditinfo  = new Button({
+            label: "Edit info",
+            added_class: ["options__button", "options__button__editinfo"],
+            type: "button",
+            shape_dir: "left",
+            bgshape: true,
+            events: {
+                click: () => { 
+                    animateClick(this.children.buttoneditinfo.element);
+                    setTimeout(() =>  {
+                        for(const profile of document.querySelectorAll<HTMLElement>('.options__profile'))
+                            profile.style.display = 'none';
+                        for(const profile of document.querySelectorAll<HTMLElement>('.options__editinfo'))
+                            profile.style.display = 'block';
+                    }, 400);
+                }
+            }
+          });
+
+        this.children.buttonavatarupload = new Button({
+            label: "Submit",
+            added_class: "",
+            type: "button",
+            events: {
+              click: () => { 
+                animateClick(this.children.buttonavatarupload.element);
+                setTimeout(() =>  {
+                  const avatarupload: HTMLElement | null = document.querySelector(".avatarupload")
+                  if(avatarupload) {
+                    avatarupload.style.opacity = '0';
+                    setTimeout(() =>  {
+                      avatarupload.style.display = 'none';
+                    }, 200);
+                  }
+                }, 400);
+              }
+            }
+        });
+
+        this.children.buttonsavepassword = new Button({
+            label: "Save",
+            added_class: ["options__button", "options__button__savepassword"],
+            type: "button",
+            bgshape: true,
+            events: {
+              click: () => {
+                animateClick(this.children.buttonsavepassword.element);
+                setTimeout(() =>  {
+                  for(const editpassword of document.querySelectorAll<HTMLElement>('.options__editpassword'))
+                      editpassword.style.display = 'none';
+                  for(const profile of document.querySelectorAll<HTMLElement>('.options__profile'))
+                      profile.style.display = 'block';
+                }, 400);
+              }
+            }
+        });
+
+        this.children.buttonsaveinfo = new Button({
+            label: "Save",
+            added_class: ["options__button", "options__button__saveinfo"],
+            type: "button",
+            bgshape: true,
+            events: {
+              click: () => { 
+                animateClick(this.children.buttonsaveinfo.element);
+                setTimeout(() =>  {
+                  for(const editinfo of document.querySelectorAll<HTMLElement>('.options__editinfo'))
+                      editinfo.style.display = 'none';
+                  for(const profile of document.querySelectorAll<HTMLElement>('.options__profile'))
+                      profile.style.display = 'block';
+                }, 400);
+              }
+            }
+        });
+
         this.childrenCollection.titles = this.props.titles.map((title: OptionsRowProps) => new OptionsRow(title))
         this.props.titles = [];
         this.childrenCollection.titles__editinfo = this.props.titles__editinfo.map((title: OptionsRowProps) => new OptionsRow(title))
