+++ src/utils/EventBus.ts
@@ -23,7 +23,7 @@ export class EventBus<
       throw new Error(`Нет события: ${event}`);
     }
 
-    this.listeners[event] = this.listeners[event]!.filter(
+    this.listeners[event] = this.listeners[event]?.filter(
       listener => listener !== callback
     );
   }
@@ -33,7 +33,7 @@ export class EventBus<
       throw new Event(`Нет события: ${event}`);
     }
 
-    this.listeners[event]!.forEach(listener => {
+    this.listeners[event]?.forEach(listener => {
       listener(...args);
     });
   }
