+++ src/utils/hasClass.ts
@@ -1,4 +1,4 @@
-export default function hasClass(el: Element, class_to_check: string) {
+export default function hasClass(el: HTMLElement | Element, class_to_check: string) {
     if(!el || !el.classList)
         return false
     return el.classList.contains(class_to_check)
